+++ history.md
@@ -5,6 +5,7 @@
     * Update: error handling when arguments aren't specified within the utils
     * Fix: static video in byte range would always stay Content-Type=video/mp4
     * Update: response functions throw error if insufficent arguments
+    * Add: response function res.sendStatus(statusCode)
 
 1.0.0 / 2022-01-13
 ===================